==> 1 find protagonists.ipynb <==
import os
import requests
import time
import pandas as pd
import openai
client = openai.OpenAI()

kaggle_data = pd.read_csv("data/kaggle_data.csv")

kaggle_data.columns

if os.path.exists("data/intermediate_protagonists.csv"):
    protagonist_map = pd.read_csv("data/intermediate_protagonists.csv", index_col=0).to_dict()["0"]
    protagonist_map = {str(k).zfill(7): v for k, v in protagonist_map.items()}

else:
    protagonist_map = {}

protagonist_map

for i in range(kaggle_data.shape[0]):
    imdb_id = kaggle_data.iloc[i]["imdb_id"]
    title = kaggle_data.iloc[i]["title"]
    if str(imdb_id).zfill(7) in protagonist_map:
        print(f"Skipping movie {title} as protagonist already found.")
        continue
    
    description = kaggle_data.iloc[i]["plot"]
    director = kaggle_data.iloc[i]["director"]

    prompt = f"""Search the internet for the protagonist of the following movie.
    Title: {title}
    Description: {description}
    Director: {director}
    Never write anything other than the name of the protagonist.
    No explanations, no full sentences, no references--just the name of the protagonist.
    If there is no clear protagonist, return an empty string.
    Protagonist name:"""

    response = client.responses.create(
    model="gpt-4.1-mini",
    tools=[{"type": "web_search"}],
    input = prompt,
    temperature=0)
    answer = response.output_text

    #write raw output to protagonist map
    protagonist_map[str(imdb_id).zfill(7)] = answer
    if i % 10 == 0:
        print(i)
        pd.Series(protagonist_map).to_csv(f"data/intermediate_protagonists.csv")
pd.Series(protagonist_map).to_csv(f"data/intermediate_protagonists.csv")

pd.Series(protagonist_map).to_csv(f"data/intermediate_protagonists.csv")

#add protagonist from dict to kaggle_data in new column
kaggle_data["character_name_search"] = kaggle_data["imdb_id"].apply(lambda x: protagonist_map.get(str(x).zfill(7), ""))
kaggle_data.to_csv("data/kaggle_data.csv", index=False)
